+++ Changelog.md
@@ -1,2 +1,20 @@
+### v 0.2.0:
+- Separate resource definitions support.
+- Bug fixes.
+
+### v 0.1.0:
+- Added support for custom actor parameters.
+
+### v 0.0.3:
+- Bug fixes.
+
+### v 0.0.2:
+- Added `additionalRequires` actor system option, which allows requiring additional
+modules in forked process.
+- Added module-based marshaller support.
+- Added `Actor.forwardToChild()` method.
+- Fixed variable argument messages for forked mode.
+- Added `Actor.metrics()` method and the metrics facility.
+
 ### v 0.0.1:
 - Initial import from SAYMON project with some necessary corrections.